+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Add `getCentroid` method to features.
+
+
 ## [0.10.0] - 2018-11-02
 
 ### Added